Use user's home directory as datadir in snap
Snaps store their data in a confined directory, this directory is
removed when you uninstall the snap. While this is a clean and safe (no
write access outside this confined space) design it's not the best
option for a cryptocurrency wallet. If the user accidentially removes
the snap it will remove their private keys as well. In other words: it's
too easy to loose your most important data.

This commit creates a `reden` directory in the user's home directory and
uses that as the wallet's datadir.